Refactor app/ entry point for clarity
