+++ client/src/app/app.tsx
@@ -1,0 +1,24 @@
+import { useAuth } from "@clerk/clerk-react"
+import { RouterProvider } from "@tanstack/react-router"
+
+import Provider from "./provider"
+import { router } from "./router"
+
+function InnerApp() {
+    const auth = useAuth()
+
+    return (
+        <RouterProvider
+            router={router}
+            context={{ auth }}
+        />
+    )
+}
+
+export default function App() {
+    return (
+        <Provider>
+            <InnerApp />
+        </Provider>
+    )
+}
